# api.yaml
openapi: 3.0.3
info:
  title: UniTrivia
  description: API to the game of UniTrivia
  version: 1.0.0
  license:
    name: MIT License
    url: https://opensource.org/licenses/mit-license.php
servers:
  - url: http://localhost:3000/api
paths:
  /login:
    get:
      tags:
        - Usuario
      description: |
        Se le pasa como parámetros el nombre de usuario y la contraseña en texto plano.

        Devuelve el identificador temporal con el que identificar todas las operaciones
        relacionadas con el usuario y un error en caso de no poder identificarlo.
      operationId: login
      parameters:
        - name: username
          in: header
          required: true
          schema:
            type: string
        - name: password
          in: header
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Login realizado exitosamente
          content:
            application/json:
              schema:
                type: string
        default:
          description: Error en el login
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /register:
    post:
      tags:
        - Usuario
      description: |
        Se le pasa comos parámetros el nombre de usuario, la contraseña en texto plano, el email
        la pregunta de seguridad y la respuesta a dicha presgunta.

        Devuelve un mensaje de confirmación de caso de poder crear el usuario y un error en caso de no poder crearlo.
      operationId: register
      parameters:
        - name: username
          in: header
          required: true
          schema:
            type: string
        - name: password
          in: header
          required: true
          schema:
            type: string
        - name: email
          in: header
          required: true
          schema:
            type: string
        - name: preg
          in: header
          required: true
          schema:
            type: string
        - name: res
          in: header
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Usuario creado exitosamente
          content:
            application/json:
              schema:
                type: string
        default:
          description: Error en el registro
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /profile:
    get:
      tags:
        - Usuario
      description: |
        Obtiene toda la información del usuario: nombre de usuario, email, pregunta de seguridad,
          respuesta de seguridad, lista de objectos comprados, estadísticas del jugador, opciones
          de personalización y cantidad de monedas.
      security:
        - jwt: [ ]
      parameters:
        - name: jwt
          in: header
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Objeto con toda la información del usuario.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Usuario'
        default:
          description: Error al obtener el usuario
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      tags:
        - Usuario
      description: |
        Elimina la información del usuario de la base de datos
      security:
        - jwt: [ ]
      parameters:
        - name: jwt
          in: header
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Respuesta de exito
          content:
            application/json:
              schema:
                type: string
        default:
          description: Error al eliminar el usuario
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  schemas:
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: integer
          format: int32
        message:
          type: string
    Usuario:
      type: object
      properties:
        username:
          description: Nombre de usuario
          type: string
        email:
          description: Correo Electrónico del usuario
          type: string
        preg:
          description: Pregunta de seguridad del usuario
          type: string
        res:
          description: Respuesta de seguridad del usuario
          type: string
        coins:
          description: Cantidad de monedas del usuario
          type: integer
        nJugadas:
          description: Número de partidas jugadas por el usuario
          type: integer
        nGanadas:
          description: Número de partidas ganadas por el usuario
          type: integer
        avatar:
          description: El ObjectID del avatar del usuario
          type: string
        banner:
          description: El ObjectID del banner del usuario
          type: string
        ficha:
          description: El ObjectID de la forma de la ficha del usuario
          type: string
        comprados:
          description: El array de ObjectID con los objetos comproados por el usuario del usuario
          type: array
          items:
            type: string
  securitySchemes:
    jwt: # arbitrary name for the security scheme
      type: apiKey
      in: header
      name: jwt